==> Notebook/14_all_combination_GE.Rmd <==
Jai Babe Di
Jai Guru Maa Ji
---
title: "R Notebook"
output: html_notebook
---

```{r}
library(nanostringr)
library(ComplexHeatmap)
library(dplyr)
library(ggplot2)
library(circlize)
library(DESeq2)
library(readxl)
library(reshape2)
library(msigdbr)
library(fgsea)
library(clusterProfiler)
library(enrichplot)
set.seed(123456)
setwd("~/Desktop/Nanostring_mRNA_project/")
```




```{r}
# load normalized samples files
liver <- read.csv("~/Desktop/Nanostring_mRNA_project/output_files/normalized_liver_data.csv", check.names = FALSE)
colnames(liver) <- str_replace_all(colnames(liver),"\\s+","_")
rownames(liver)<- c(liver[1])[[1]]
liver <- liver[-1]
colnames(liver)[colnames(liver) == "0621_BLOCK_7_16_2206_7"] <- "0621_BLOCK_7_16_2206_7_liver"
################################
lungs <- read.csv("~/Desktop/Nanostring_mRNA_project/output_files/normalized_lung_data.csv", check.names = FALSE)
colnames(lungs) <- str_replace_all(colnames(lungs),"\\s+","_")
rownames(lungs)<- c(lungs[1])[[1]]
lungs <- lungs[-1]
################################
kidney <- read.csv("~/Desktop/Nanostring_mRNA_project/output_files/normalized_kidney_data.csv", check.names = FALSE)
rownames(kidney)<- c(kidney)[[1]]
kidney <- kidney[-1]
colnames(kidney) <- str_replace_all(colnames(kidney),"\\s+","_")
colnames(kidney)[colnames(kidney) == "0621_BLOCK_7_16_2206_7"] <- "0621_BLOCK_7_16_2206_7_kid"
```

```{r}
combined_all <- cbind(liver,lungs, kidney)
colnames(combined_all)<- toupper(colnames(combined_all))

write.csv(combined_all, "~/Desktop/Nanostring_mRNA_project/output_files/normalized_all_data.csv")
```


```{r}
# pathway for GROUP 2
meta_data <- read_excel("~/Desktop/Nanostring_mRNA_project/NATanalysis1_29july.xlsx",sheet = "All")
#meta_data<- meta_data[!meta_data$`NAT Study - met to this tissue?` %in% c("yes - distant (FFPE)","yes - proximal/adjacent (FFPE)"),]
#meta_data<- meta_data[!meta_data$GROUPS_1 %in% "include or not?",]
meta_data$`File name`<- toupper(meta_data$`File name`)
meta_data<- as.data.frame(meta_data)
rownames(meta_data)<- meta_data$`File name`
meta_data$`File name` <- str_replace_all(meta_data$`File name`,"\\s+","_")

meta_data <- meta_data[meta_data$`File name` %in% colnames(combined_all),]
# colnames(combined_all)[!colnames(combined_all) %in% meta_data$`File name`]
```


```{r}
# remove entries include or not in x2 also
combined_all <- combined_all[, colnames(combined_all) %in% meta_data$`File name`]

# Create a data frame for the annotations
sample_order <- match(colnames(combined_all), meta_data$`File name`)
metadata_ordered <- meta_data[sample_order, ]
```




```{r}
annotations <- data.frame(
   GROUPS_1 = metadata_ordered$GROUPS_1,
  GROUPS_2 = metadata_ordered$`NAT Tissue`,
  row.names = metadata_ordered$`File name`
)

# Create the heatmap annotations
ha <- rowAnnotation(
  GROUPS_2 = annotations$GROUPS_2,
  GROUPS_1 = annotations$GROUPS_1,
  col = list(
     GROUPS_2 = c( "Liver" = "#0072B2", "Lung" = "#999999","Kidney" = "#D6604D" ),
    GROUPS_1 = c("Control" = "#E69F00", "No-met" = "#009E73", "Yes-met" = "#CC79A7" )
    
    ))

# Define custom colors for the heatmap
heatmap_colors <- colorRamp2(c(-4, 0, 4), rev(RColorBrewer::brewer.pal(11, "RdBu")[c(2, 6, 11)]))

#pdf("~/Desktop/Nanostring_mRNA_project/Plots/combined_all_expression_heatmap.pdf", height = 4, width = 4)

#colnames(x2) <- sub("^(([^ ]* ){2}[^ ]*) .*$", "\\1", colnames(x2))
# Create and draw the heatmap
Heatmap(
  scale(t(combined_all)),
  name = "Expression",
  left_annotation = ha,
  show_row_names = F,
  show_column_names = F,
  col = heatmap_colors,
  cluster_rows = F,
  cluster_columns = T,
  show_column_dend = F,
  show_row_dend = F,
  #show_parent_dend_line = FALSE,
  row_title_gp = gpar(fontsize = 8,fontface = "bold"),
  column_title_gp = gpar(fontsize = 8,  fontface = "bold"),
#  column_names_gp = gpar(fontsize = 6, angle = 90, vjust = 1, hjust = 1),
  row_names_gp = gpar(fontsize = 6, fontface = "bold"),
  row_split = annotations$GROUPS_1,
  row_title = "Sample ID"
)
```

```{r}

res <- cor(combined_all)
col<- colorRampPalette(c("blue", "white", "red"))(20)

pdf("~/Desktop/Nanostring_mRNA_project/Plots/combined_all_correlation_heatmap.pdf", height = 10, width = 12)
Heatmap(
  res,
  name = "Pearson R",
  left_annotation = ha,
  show_row_names = T,
  show_column_names =T,
  col = col,
  cluster_rows = T,
  cluster_columns = T,
  show_column_dend = F,
  show_row_dend = F,
  #show_parent_dend_line = FALSE,
  row_title_gp = gpar(fontsize = 8,fontface = "bold"),
  column_title_gp = gpar(fontsize = 8,  fontface = "bold"),
  column_names_gp = gpar(fontsize = 8, angle = 45, vjust = 1, hjust = 1, fontface = "bold"),
  row_names_gp = gpar(fontsize = 8, fontface = "bold"),
  row_split = annotations$GROUPS_2,
 row_title = NULL
)

```


```{r}
# remove entries include or not in x2 also
combined_all <- combined_all[, colnames(combined_all) %in% meta_data$`File name`[meta_data$GROUPS_1 =="Yes-met"]]

sample_order <- match(colnames(combined_all), meta_data$`File name`)
metadata_ordered <- meta_data[sample_order, ]

annotations <- data.frame(
   GROUPS_1 = metadata_ordered$GROUPS_1,
  GROUPS_2 = metadata_ordered$`NAT Tissue`,
  row.names = metadata_ordered$`File name`
)

annotations$GROUPS_2 <- factor(annotations$GROUPS_2, levels = c("Liver","Lung","Kidney"))

# Create the heatmap annotations
ha2 <- rowAnnotation(
  GROUPS_2 = annotations$GROUPS_2,
  col = list(
     GROUPS_2 = c("Lung" = "#999999","Liver" = "#0072B2", "Kidney" = "#D6604D" )
    
    ))

res <- cor(combined_all)
col<- colorRampPalette(c("blue", "white", "red"))(20)

pdf("~/Desktop/Nanostring_mRNA_project/Plots/combined_all_yes_met_correlation_heatmap.pdf", height = 10, width = 12)
Heatmap(
  res,
  name = "Pearson R",
  left_annotation = ha2,
  show_row_names = T,
  show_column_names =T,
  col = col,
  cluster_rows = F,
  cluster_columns = F,
  show_column_dend = F,
  show_row_dend = F,
  #show_parent_dend_line = FALSE,
  row_title_gp = gpar(fontsize = 10,fontface = "bold"),
  column_title_gp = gpar(fontsize = 10,  fontface = "bold"),
  column_names_gp = gpar(fontsize = 10, angle = 45, vjust = 1, hjust = 1, fontface = "bold"),
  row_names_gp = gpar(fontsize = 10, fontface = "bold"),
  row_split = annotations$GROUPS_2,
 row_title = NULL
)
```
